rfc: Implement fmt::Debug for vcpu and exit context
This change implements fmt::Debug for vcpu, exit context and
a few other related structures, providing detailed and readable
information.

Not only are we printing the structure fields, but we're also
using WHP functions to fetch vcpu registers and counters as well
as parsing unions. This can be really convenient for testing
and debugging purposes.

Co-authored-by: Timo Kreuzer <timo@crowdstrike.com>