Pull vmm-vcpu trait implementations and structures
The vmm-vcpu crate provides an abstraction layer for libwhp, making
it easy to switch between KVM and WHP.

This is a bit ouf of the scope of this crate so we're moving it
to the already existing "whp" crate, which may become a rust-vmm
crate in the future.

The example using those traits has also been moved to the whp crate.